Tlsfix

+++ Connection.js
@@ -1,47 +1,94 @@
 const net = require('net');
 const tls = require('tls');
 const Settings = require('./settings');
+const { ScreenUpdate } = require('./Helpers/ScreenUpdate');
+const { Listener } = require('./Listener');
 
-class Connecting {
-  constructor(server, port, user, realname, nickname, channels, tlsFlag) {
-    this.nickname = Settings[this.identity]?.nickname || nickname || Settings.nicknames[0];
+class Connection extends Listener {
+  constructor({server, port, user, realname, nickname, channels, password, serverpassword, tls, rejectUnauthorized}) {
+    super();
+    this.originalNick = nickname;
+    this.nickname = nickname || Settings[this.identity]?.nickname  || Settings.nicknames[0];
     this.server = server;
     this.port = port;
     this.channels = channels;
     this.user = user || Settings.USER;
     this.realname = realname || Settings.realname;
     this.connecting = false;
-    this.tlsFlag = tlsFlag;
-    this.MakeIdentity();
-    this.CreateSettings();
+    this.preError = false;
+    this.tlsFlag = tls;
+    this.password = password;
+    this.serverpassword = serverpassword;
+    this.identity;
     this.options = {
-      rejectUnauthorized: false
+      rejectUnauthorized: rejectUnauthorized,
+      port: this.port,
+      host: this.server
     };
   }
 
+  Start() {
+    this.MakeIdentity();
+    this.CreateSettings();
+    this.Connect();
+  }
+
+  MakeIdentity() {
+    this.identity = new Date().getTime() + Math.floor(Math.random() * 100);
+  }
+
+  CreateSettings() {
+    Settings[this.identity] = {
+      private: {},
+      identity: this.identity,
+      nickname: this.nickname,
+      server: this.server,
+      port: this.port,
+      channels: this.channels,
+      user: this.user,
+      realname: this.realname,
+      PassedMOTD: false,
+      count: 0,
+      joinedChans: [],
+      displayRaw : false,
+      raw : "",
+      status : ""
+    };
+  }
+
+  RebindClient(client){
+    this.client = client;
+  }
 
   Connect() {
-    const client = this.tlsFlag ? new tls.TLSSocket() : new net.Socket();
-    this.client = client;
-    this.preError = true;
-    client.setTimeout(80000);
+    let client = null;
+    const { Listen, identity , server, tlsFlag, user, realname, serverpassword, password, originalNick } = this;
     const connect = () => {
-      client.connect(this.port, this.server, this.options);
+      if(this.tlsFlag){
+        client = tls.connect(this.options);
+      } else {
+        client = net.connect(this.options); 
+      }
       client.setTimeout(80000);
       this.connecting = false;
-      Settings[this.identity].PassedMOTD = false;
-      clearTimeout(Settings[this.identity].timeout);
-      clearTimeout(Settings[this.identity].ping);
-      clearTimeout(Settings[this.identity].connect);
+      Settings[identity].PassedMOTD = false;
+      clearTimeout(Settings[identity].ping);
+      clearTimeout(Settings[identity].connect);
       this.preError = false;
+      BindEvents(client);
+      this.RebindClient(client);
+      ScreenUpdate({identity, client, server});
+      this.UpdateScreen = true;
     };
 
     const connectHandler = () => {
+      let nickname = Settings[this.identity]?.nickname || this.nickname;
       client.write('CAP LS 302\r\n');
       client.write('CAP REQ multi-prefix\r\n');
       client.write('CAP END\r\n');
-      client.write(`NICK ${this.nickname}\r\n`);
-      client.write(`USER ${this.user} 0 * :${this.realname}\r\n`);
+      serverpassword ? client.write(`PASS ${serverpassword}\r\n`) : null;
+      client.write(`NICK ${nickname}\r\n`);
+      client.write(`USER ${user} 0 * :${realname}\r\n`);
       this.connecting = true;
       this.preError = false;
     };
@@ -54,12 +101,16 @@ class Connecting {
           Settings[this.identity][chan.toLowerCase()].chanNicks = [];
         });
       }
-      client.destroy();
+      if (client) {
+        client.destroy();
+        client = null;
+      }
       Settings[this.identity].connect = setTimeout(connect, 5000);
       Settings[this.identity].status += `^R${err} Reconnecting...^\r\n`;
     };
 
     const errorHandler = (err) => {
+      console.logger(err);
       this.connecting = false;
       if (!this.preError) {
         this.preError = true;
@@ -85,36 +136,20 @@ class Connecting {
       }
     };
 
-    client.on('error', errorHandler);
-    client.on('connect', connectHandler);
-    client.on('timeout', Timeout);
-    client.on('close', Close);
-   
+    function BindEvents(client){
+      client.on('error', errorHandler);
+      tlsFlag ? client.on('secureConnect', connectHandler) :
+      client.on('connect', connectHandler);
+      client.on('timeout', Timeout);
+      client.on('close', Close);
+      client.on("data", (data) => Listen(data, identity, client, server, password, originalNick));
+      
+    }
+
     connect();
   }
 
-  CreateSettings() {
-    Settings[this.identity] = {
-      private: {},
-      identity: this.identity,
-      nickname: this.nickname,
-      server: this.server,
-      port: this.port,
-      channels: this.channels,
-      user: this.user,
-      realname: this.realname,
-      PassedMOTD: false,
-      count: 0,
-      joinedChans: [],
-      displayRaw : false,
-      raw : "",
-      status : ""
-    };
-  }
 
-  MakeIdentity() {
-    this.identity = new Date().getTime() + Math.floor(Math.random() * 100);
-  }
 }
 
-module.exports = { Connecting };
+module.exports = { Connection };
